+++ include/private/common/compat.hpp
@@ -14,5 +14,27 @@ namespace mk {
 template <typename T> using Var = SharedPtr<T>;
 #endif
 
+#if MK_VERSION_MAJOR > 0 || (MK_VERSION_MINOR < 9)
+/*
+Simplifies life when you use templates for mocking APIs because
+allows you to write the following:
+
+    template <MK_MOCK(event_base_new)>
+    void foobar() {
+        event_base *p = event_base_new();
+    }
+
+Which is arguably faster than writing the following:
+
+    template <decltype(event_base_new) event_base_new = ::event_base_new>
+    void foobar() { ...
+*/
+#define MK_MOCK(name_) MK_MOCK_AS(name_, name_)
+
+// Similar to MK_MOCK but with alias
+#define MK_MOCK_AS(name_, alias_) decltype(name_) alias_ = name_
+
+#endif // ...
+
 } // namespace mk
 #endif
